requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-redis
qrcode
reportlab
mysqlclient
mysql-connector-python
opencv-python-headless
hashlib
xhtml2pdf
